Move world graphics logic into gfx_world.cpp
