+++ CMakeLists.txt
@@ -664,6 +664,7 @@ set(THEXTECH_SRC
     src/graphics/gfx_editor.cpp
     src/graphics/gfx_frame.cpp
     src/graphics/gfx_camera.cpp
+    src/graphics/gfx_world.cpp
     src/graphics/gfx_credits.cpp
     src/graphics/gfx_hud.cpp
     src/graphics/gfx_enter_screen.cpp
